pyproject.toml
[project]
name = "dbserver-api"
version = "0.1.0"
description = "DB Server API"
authors = [
    {name = "thehecktour",email = "thehacktour@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)",
    "django-stubs (>=5.2.7,<6.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]
